routing_service/requirements.txt
alembic==1.14.0
fastapi==0.115.6
pydantic==2.10.3
SQLAlchemy==2.0.36
uvicorn==0.32.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-mock==3.14.0
black==24.10.0
sentry-sdk[fastapi]==2.19.2
coverage==7.6.9
